# part1/artist-api-spec.yaml
openapi: 3.1.1
info:
  title: Record Label API
  version: 1.0.0
  description: |
    OpenAPI 3.1.1 specification for a record label artist service.
    The API is secured with Basic Authentication and allows clients to:
    - list artists with pagination using `offset` and `limit`
    - create a new artist
    - fetch one artist using the `artistname` path parameter
    
    Design rationale:
    - OpenAPI 3.1.1 is used because it is the current OpenAPI format and supports schema-first API design.
    - Basic Authentication is suitable here because the assignment explicitly requires simple authenticated access for authorized clients.
    - Pagination is included to avoid returning the entire artist collection in one response and to improve scalability.
    - The `offset` parameter is treated as a page number and `limit` is the page size, as requested in the assignment text.

servers:
  - url: https://api.recordlabel.com
    description: Production server

security:
  - basicAuth: []

tags:
  - name: Artists
    description: Artist operations

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Only authorized users can consume this API.
        Send credentials in the `Authorization` header using HTTP Basic Authentication.

  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Page number to return, starting from 0. In this specification, `offset` represents the page index rather than the raw row displacement.
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0

    LimitParam:
      name: limit
      in: query
      description: Number of artist records to return on one page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10

    ArtistNameParam:
      name: artistname
      in: path
      required: true
      description: Artist name used to fetch one artist record. This follows the assignment wording, which explicitly requests an `artistname` path parameter.
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: Taylor Swift

  schemas:
    Artist:
      type: object
      required:
        - artistName
        - artistGenre
        - albumsPublished
        - artistUsername
      properties:
        artistName:
          type: string
          minLength: 1
          maxLength: 100
          example: Taylor Swift
        artistGenre:
          type: string
          minLength: 1
          maxLength: 50
          example: Pop
        albumsPublished:
          type: integer
          minimum: 0
          example: 10
        artistUsername:
          type: string
          pattern: '^[A-Za-z0-9_]+$'
          minLength: 3
          maxLength: 30
          example: taylorswift

    ArtistInput:
      type: object
      required:
        - artistName
        - artistGenre
        - albumsPublished
        - artistUsername
      properties:
        artistName:
          type: string
          minLength: 1
          maxLength: 100
          example: Adele
        artistGenre:
          type: string
          minLength: 1
          maxLength: 50
          example: Soul
        albumsPublished:
          type: integer
          minimum: 0
          example: 4
        artistUsername:
          type: string
          pattern: '^[A-Za-z0-9_]+$'
          minLength: 3
          maxLength: 30
          example: adele

    ArtistListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        pagination:
          type: object
          required:
            - offset
            - limit
            - totalRecords
          properties:
            offset:
              type: integer
              example: 0
            limit:
              type: integer
              example: 10
            totalRecords:
              type: integer
              example: 25

    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized

  responses:
    BadRequest:
      description: Invalid request data or invalid pagination values.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            message: Invalid request parameters

    Unauthorized:
      description: Missing or invalid Basic Authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            message: Unauthorized

    NotFound:
      description: The requested artist was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            message: Artist not found

    Conflict:
      description: Artist already exists in the database.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 409
            message: Artist username already exists

    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 500
            message: Internal server error

paths:
  /artists:
    get:
      tags:
        - Artists
      summary: Get a list of artists
      description: |
        Returns a paginated list of artists using `offset` and `limit`.
        A `200` response is used for successful retrieval, `400` for invalid pagination input,
        `401` for authentication failure, and `500` for unexpected server-side errors.
      operationId: getArtists
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Artist list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistListResponse'
              example:
                data:
                  - artistName: Taylor Swift
                    artistGenre: Pop
                    albumsPublished: 10
                    artistUsername: taylorswift
                  - artistName: Ed Sheeran
                    artistGenre: Pop
                    albumsPublished: 6
                    artistUsername: edsheeran
                pagination:
                  offset: 0
                  limit: 10
                  totalRecords: 25
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Artists
      summary: Add a new artist
      description: |
        Creates a new artist in the record label database.
        A `201` response is used because a new resource is created, `400` indicates invalid input,
        `401` indicates unauthorized access, `409` indicates a uniqueness conflict, and `500`
        represents an unexpected server error.
      operationId: createArtist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistInput'
      responses:
        '201':
          description: Artist created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
              example:
                artistName: Adele
                artistGenre: Soul
                albumsPublished: 4
                artistUsername: adele
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /artists/{artistname}:
    get:
      tags:
        - Artists
      summary: Get a specific artist
      description: |
        Returns a single artist record based on the `artistname` path parameter.
        A `404` response is returned when the named artist does not exist in the collection.
      operationId: getArtistByName
      parameters:
        - $ref: '#/components/parameters/ArtistNameParam'
      responses:
        '200':
          description: Artist returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
              example:
                artistName: Taylor Swift
                artistGenre: Pop
                albumsPublished: 10
                artistUsername: taylorswift
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'